refactor: pull members up from JavaBasePipelineExt to AbstractPipelineExt
